MU-195 Fix lagging when navigate to chat
