+++ src/src/screens/Groups/Workspace/index.tsx
@@ -30,7 +30,6 @@ import { useMobxStore } from "~/mobx/store";
 const Workspace: ScreenProps<"workspace"> = ({ route }) => {
   // Needed data
   const groupId = route.params.groupId;
-  // const currentUser = useAppSelector(state => state.user.data);
   const navigation = useNavigation();
   const {
     data: workspace,
